PIM-7211: Convert header to typescript
