+++ package.json
@@ -39,6 +39,7 @@
     "@types/backbone": "^1.3.42",
     "@types/node": "^9.6.1",
     "JSON2": "0.1.0",
+    "add-module-exports-webpack-plugin": "^0.1.0",
     "babel-cli": "6.26.0",
     "babel-core": "6.26.0",
     "babel-eslint": "7.2.3",
