+++ src/Pim/Bundle/EnrichBundle/Resources/config/requirejs.yml
@@ -642,7 +642,7 @@ config:
 
         # Catalog volume monitoring
         pim/catalog-volume/section:                      pimenrich/js/catalog-volume/section
-        pim/catalog-volume/header:                       pimenrich/js/catalog-volume/header
+        pim/catalog-volume/header:                       pimenrich/js/catalog-volume/header.ts
 
         # Channel
         pim/channel-edit-form/delete:                           pimenrich/js/channel/form/delete
